requirements.txt
fastapi==0.120.2
uvicorn==0.38.0
sqlalchemy==2.0.44
asyncpg==0.30.0
alembic==1.17.1
passlib==1.7.4
pydantic-settings==2.11.0
dnspython==2.8.0 
email-validator==2.3.0
pyjwt==2.10.1
python-multipart==0.0.20